+++ package.json
@@ -9,7 +9,7 @@
     "dev": "NODE_ENV=development concurrently --kill-others-on-fail npm:dev-ui",
     "build": "NODE_ENV=production npm run build-ui",
     "build:ci": "NODE_ENV=development npm run build-ui",
-    "test": "NODE_ENV=test jest --coverage=false --silent",
+    "test": "NODE_ENV=test jest --config jest.config.ts --coverage=false --silent",
     "test:ci": "NODE_ENV=test jest --silent --coverage=false --reporters=jest-silent-reporter",
     "test:coverage": "NODE_ENV=test jest --coverage",
     "format:check": "prettier --check 'src'",
@@ -87,26 +87,26 @@
     "@types/redux-logger": "^3.0.9",
     "@types/sinon": "^10.0.13",
     "@types/webtorrent": "^0.109.3",
+    "@typescript-eslint/eslint-plugin": "^5.43.0",
+    "@typescript-eslint/parser": "^5.43.0",
     "assert": "^2.0.0",
     "browserify": "^17.0.0",
     "concurrently": "^7.6.0",
+    "constants-browserify": "^1.0.0",
     "copy-webpack-plugin": "^11.0.0",
     "crypto-browserify": "^3.12.0",
     "css-loader": "^6.7.2",
-    "fake-indexeddb": "^4.0.0",
-    "file-loader": "^6.2.0",
-    "html-entities": "^2.3.3",
-    "html-webpack-plugin": "^5.5.0",
-    "https-browserify": "^1.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.43.0",
-    "@typescript-eslint/parser": "^5.43.0",
-    "constants-browserify": "^1.0.0",
     "eslint": "^5.16.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-import-resolver-typescript": "^3.5.2",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-n": "^15.5.1",
     "eslint-plugin-promise": "^6.1.1",
+    "fake-indexeddb": "^4.0.0",
+    "file-loader": "^6.2.0",
+    "html-entities": "^2.3.3",
+    "html-webpack-plugin": "^5.5.0",
+    "https-browserify": "^1.0.0",
     "husky": "^8.0.2",
     "image-webpack-loader": "^8.1.0",
     "jest": "^29.3.1",
@@ -125,6 +125,7 @@
     "style-loader": "^3.3.1",
     "ts-jest": "^29.0.3",
     "ts-loader": "^9.4.1",
+    "ts-node": "^10.9.1",
     "typescript": "^4.9.3",
     "webpack": "^5.75.0",
     "webpack-cli": "^5.0.0",
